requirements.txt
flask
python-jose 
requests 
cachetools
python-keycloak